feat: add Duration type path to the CypherFilter
