+++ packages/graphql/src/translate/queryAST/ast/filters/utils/create-point-operation.ts
@@ -18,8 +18,8 @@
  */
 
 import Cypher from "@neo4j/cypher-builder";
-import type { AttributeAdapter } from "../../../schema-model/attribute/model-adapters/AttributeAdapter";
-import type { WhereOperator } from "../ast/filters/Filter";
+import type { AttributeAdapter } from "../../../../../schema-model/attribute/model-adapters/AttributeAdapter";
+import type { WhereOperator } from "../Filter";
 
 export function createPointOperation({
     operator,
